ipfixprobe: update to version 2.18.0 and Makefile polishing
- Removed OpenWrt license as we are not affiliated anyhow with OpenWrt.

- These days PKG_REV is not commonly used in OpenWrt packages feed.
So, let's get rid of it. It was replaced by using PKG_SOURCE_DATE, PKG_SOURCE_VERSION
Also, it simplifies Makefile. There is also possibility to use tags or
releases, but in ipfixprobe, there are none.

- Value in variable PKG_BUILD_DIR is default.
This can be checked in file include/package.mk from OpenWrt main
repository. Let's remove it from this Makefile.

- Corrected License as ipfixprobe is using BSD-3-Clause or GPL-2.0-or-later
by checking individual files.
I am using valid SPDX License Identifier.

- This is NEMEA module, so it should be listed under NEMEA submenu.

- Added description of this package.
This can be seen in make menuconfig
and on the router while using opkg info and/or opkg list

- Removed dummy MENU in make menuconfig caused by MENU=1
(screenshots can be found in mine pull request)

- Fixed order in conffile define.
It means that config file is marked as configuration file and it won't be
overwritten by an update, which is now done by an incorrect order.

- Removed define Build/Configure as we are calling default.
It works without it.

- Changed permission of ipfixprobe config from 0644 to 0600

Compile and run tested on Turris Omnia, Turris OS 5.1.3 based on OpenWrt 19.07.4

Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>